requirements.txt
mysql-connector-python 
colorama 
termcolor